+++ pallets/subbridge/src/mock/para.rs
@@ -152,7 +152,7 @@ impl pallet_parachain_info::Config for Runtime {}
 
 parameter_types! {
 	pub const TestChainId: u8 = 5;
-	pub const ResourceIdGenerationSalt: Option<u8> = Some(5);
+	pub const ResourceIdGenerationSalt: Option<u128> = Some(5);
 	pub const ProposalLifetime: u64 = 100;
 	pub const RelayNetwork: NetworkId = NetworkId::Kusama;
 	pub RelayChainOrigin: Origin = cumulus_pallet_xcm::Origin::Relay.into();
