+++ runtime/rhala/src/lib.rs
@@ -1437,7 +1437,7 @@ impl pallet_collator_selection::Config for Runtime {
 
 parameter_types! {
     pub const BridgeChainId: u8 = 1;
-    pub const ResourceIdGenerationSalt: Option<u8> = None;
+    pub const ResourceIdGenerationSalt: Option<u128> = None;
     pub const ProposalLifetime: BlockNumber = 50400; // ~7 days
     pub const BridgeEventLimit: u32 = 1024;
 }
